LayoutTests/js/dfg-post-inc-then-exit-expected.txt

Tests what happens if you post-inc and then OSR exit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS o.f is 0
PASS successfullyParsed is true

TEST COMPLETE
